t-whatpm/webidl/webidl-exception.dat

#data
exception e {
  long a;
};
#errors
1;non-module definition;s
#document
| exception e {
|   long a;
| };

#data
exception e {
  long a;
  long a;
};
#errors
1;non-module definition;s
3;duplicate member;m;a

#data
exception e {
  long a;
  long _a;
};
#errors
1;non-module definition;s
3;duplicate member;m;a

#data
exception e {
  long a;
  boolean a;
};
#errors
1;non-module definition;s
3;duplicate member;m;a

#data
exception e {
  long a;
  long b;
};
#errors
1;non-module definition;s

#data
exception e {
  long a;
  long __a;
};
#errors
1;non-module definition;s

#data
exception e {
  long a;
};
exception f {
  long a;
};
#errors
1;non-module definition;s
4;non-module definition;s

#data
[ExceptionConsts] module m {
  exception f { };
};
#errors
1;xattr id missing;m;ExceptionConsts
#document
| [ExceptionConsts] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConst] module m {
  exception f { };
};
#errors
1;unknown xattr;u;ExceptionConst
#document
| [ExceptionConst] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionsConsts] module m {
  exception f { };
};
#errors
1;unknown xattr;u;ExceptionsConsts
#document
| [ExceptionsConsts] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionsConst] module m {
  exception f { };
};
#errors
1;unknown xattr;u;ExceptionsConst
#document
| [ExceptionsConst] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConsts=f] module m {
  exception f { };
};
#errors
#document
| [ExceptionConsts=f] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConsts=_f] module m {
  exception f { };
};
#errors
#document
| [ExceptionConsts=f] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConsts=ff] module m {
  exception f { };
};
#errors
1;exception not defined;m;ff
#document
| [ExceptionConsts=ff] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConsts=m::f] module m {
  exception f { };
};
#errors
1.20;after xattr;m
#document

#data
[ExceptionConsts=::f] module m {
  exception f { };
};
#errors
1.19;before xattrarg;m
#document

#data
[ExceptionConsts=_f] module m {
  exception f { };
};
#errors
#document
| [ExceptionConsts=f] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConsts=f] module m {
  exception _f { };
};
#errors
#document
| [ExceptionConsts=f] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConsts=f] module m {
  interface f { };
};
#errors
1;exception not defined;m;f
#document
| [ExceptionConsts=f] module m {
| 
| interface f {
| };
| 
| };

#data
module m {
  exception f {};
};
[ExceptionConsts=f] module m {
};
#errors
4;duplicate qname;m;::m
4;exception not defined;m;f
#document
| module m {
| 
| exception f {
| };
| 
| };
| 
| [ExceptionConsts=f] module m {
| 
| 
| };

#data
[ExceptionConsts=f,ExceptionConsts=f] module m {
  exception f { };
};
#errors
1;duplicate xattr;w;ExceptionConsts
#document
| [ExceptionConsts=f, ExceptionConsts=f] module m {
| 
| exception f {
| };
| 
| };

#data
[ExceptionConsts=f, ExceptionConsts = g] module m {
  exception f { };
  exception g {};
};
#errors
1;duplicate xattr;w;ExceptionConsts
#document
| [ExceptionConsts=f, ExceptionConsts=g] module m {
| 
| exception f {
| };
| exception g {
| };
| 
| };

#data
[ExceptionConsts()] module m {
  exception f { };
};
#errors
1;xattr id missing;m;ExceptionConsts
1;xattr arglist not allowed;m;ExceptionConsts

#data
[ExceptionConsts=f ()] module m {
  exception f { };
};
#errors
1;xattr arglist not allowed;m;ExceptionConsts

#data
module m {
  [ExceptionConsts=m] exception f {
  };
};
#errors
2;xattr not applicable;w;ExceptionConsts

#data
module m {
  [ExceptionConsts=m] interface f {
  };
};
#errors
2;xattr not applicable;w;ExceptionConsts

#data
module m {
  interface x {
    [ExceptionConsts=m] void m ();
  };
};
#errors
3;xattr not applicable;w;ExceptionConsts

#data
module m {
exception x;
exception y{};
};
#errors
2.12;before webidl block;m;exception
#document
| module m {
| 
| exception y {
| };
| 
| };

#data
module m {
exception x {
0;
long c;
};
};
#errors
3.1;before exception member;m
#document
| module m {
| 
| exception x {
|   long c;
| };
| 
| };

#data
module m {
exception x {
[x] 0;
long a;
};
};
#errors
3.5;before exception member;m
#document
| module m {
| 
| exception x {
|   long a;
| };
| 
| };

#data
module m {
exception x {
long 0;
long a;
};
};
#errors
3.6;no webidl identifier:exception member;m
#document
| module m {
| 
| exception x {
|   long a;
| };
| 
| };
